requirements.txt
sqlalchemy
irc
pyserial
mysql-connector-python-rf
